+++ src/main/java/bumblebee/xchangepass/domain/wallet/transaction/consumer/DeadLetterConsumer.java
@@ -1,6 +1,6 @@
 package bumblebee.xchangepass.domain.wallet.transaction.consumer;
 
-import bumblebee.xchangepass.domain.wallet.wallet.dto.request.WalletTransactionMessage;
+import bumblebee.xchangepass.domain.wallet.transaction.dto.WalletTransactionMessage;
 import com.rabbitmq.client.Channel;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
